fix(billing): cast debit param so KerfPaid balance debit executes (T-141a)
_commit_paid used SQL literal VALUES ($1, -$2) — unary minus on an
untyped asyncpg parameter raises AmbiguousFunctionError (operator is
not unique: - unknown), so the KerfPaid balance debit never ran in
production (txn rolled back, usage_event not written). Cast to
-$2::numeric. Un-xfail the two T-117 guard tests — now real asserts.
test_quota.py 22 passed/0 xfailed; full kerf-billing 131 passed.

+++ packages/kerf-billing/src/kerf_billing/spend.py
@@ -181,7 +181,7 @@ async def _commit_paid(
             await conn.execute(
                 """
                 INSERT INTO cloud_user_balances (user_id, credits_usd)
-                VALUES ($1, -$2)
+                VALUES ($1, -$2::numeric)
                 ON CONFLICT (user_id) DO UPDATE
                 SET credits_usd = cloud_user_balances.credits_usd - $2
                 """,
